+++ src/RequestBuilder/RequestBuilderFactory.php
@@ -77,6 +77,11 @@ class RequestBuilderFactory
         return $this->createConfiguredBuilder(RecommendationRequestBuilder::class, $recommendation);
     }
 
+    public function resetDatabase(): ResetDatabaseRequestBuilder
+    {
+        return $this->createConfiguredBuilder(ResetDatabaseRequestBuilder::class);
+    }
+
     /**
      * @param string $builderClass
      * @param array ...$args